Update PyParse.py
Wells described only by an integer in rpt files are amended so that the user can enter a specific number of columns and rows. No longer able to reformat wells - i.e. 12 x 2 can no longer be formatted into a 6x4, as it was found this led to errors for some plate types.